feat: MIT许可证 + SMP per-CPU调度器 + 负载均衡 + RT调度
- 许可证从 GPL-3.0 切换为 MIT
- 实现 Per-CPU 运行队列 (per_cpu_rq_t)
- 实现负载均衡算法 (smp_try_balance_load)
- 添加 CPU 亲和性接口 (smp_set_affinity)
- 增强 MLFQ 调度器支持实时任务 (SCHED_FIFO/SCHED_RR)
- 添加定时优先级提升防饥饿机制 (timer.c)
- 实现阻塞式睡眠替代忙等 (timer_sleep)
- 新增 22 个 RT 调度测试用例
- 新增 14 个 SMP 测试用例
- 修复 20+ 预存语法错误

+++ src/kernel/tests/test_persistence.c
@@ -70,7 +70,7 @@ static int test_hvfs_file_persistence(void) {
     
     int written = hvfs_write(fd, test_data, strlen(test_data));
     if (written != (int)strlen(test_data)) {
-        klog_kern("[FAIL] Write failed: expected %d, got %d", strlen(test_data, written);
+        klog_kern("[FAIL] Write failed: expected %d, got %d", (int)strlen(test_data), written);
     }
     TEST_ASSERT(written == (int)strlen(test_data));
     
@@ -87,7 +87,7 @@ static int test_hvfs_file_persistence(void) {
     memset(read_buf, 0, sizeof(read_buf));
     int bytes_read = hvfs_read(fd, read_buf, sizeof(read_buf) - 1);
     if (bytes_read != (int)strlen(test_data)) {
-        klog_kern("[FAIL] Read failed: expected %d, got %d", strlen(test_data, bytes_read);
+        klog_kern("[FAIL] Read failed: expected %d, got %d", (int)strlen(test_data), bytes_read);
     }
     TEST_ASSERT(bytes_read == (int)strlen(test_data));
